+++ changelog.md
@@ -113,6 +113,7 @@ Use the RCSS `display` property to enable table formatting. See the style sheet 
 ### New RML elements
 
 - Added [Lottie plugin](https://mikke89.github.io/RmlUiDoc/pages/cpp_manual/lottie.html) for displaying vector animations using the `<lottie>` element [#134](https://github.com/mikke89/RmlUi/pull/134) (thanks @diamondhat).
+- Added `<label>` element for associating a caption with an input element.
 
 ### Element improvements